requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and Supabase
supabase==2.0.2
postgrest==0.13.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Data processing and ML
pandas==2.1.4
numpy==1.24.4
scikit-learn==1.3.2

# HTTP client for webhooks
httpx==0.25.2

# Environment variables
python-dotenv==1.0.0

# Date and time utilities
python-dateutil==2.8.2

# Production server
gunicorn==21.2.0
